New function plot.errors() that plots points with error bars
Also in pairs.custom() changed the parameter list passed to plot.binned() (when a plot.binned() is requested in one of the panels of the pairs plot) following the change of default plot in plot.binned() to a simple points plot, so that we still plot circle symbols by default (i.e. bubbles).